backend/requirements.txt
asgiref==3.5.0
Django==4.0.3
djangorestframework==3.13.1
drf-nested-routers==0.93.4
pytz==2021.3
sqlparse==0.4.2
tzdata==2021.5
